Refactor: extract VaultSearchError hierarchy to exceptions.py
schemas.py が担っていたドメイン例外定義 (VaultSearchError /
NoteNotFoundError) を新設 exceptions.py に分離。schemas.py はモデル定義に
専念させる。

- src/vault_search/exceptions.py 新規
- schemas.py: 例外定義削除
- server.py: NoteNotFoundError の import 元を .exceptions へ
- tests/test_server.py: 同上

re-export は残さない (2 通りの import path を生まない方針)。
252 tests 引き続き green、lint clean。
