+++ src/vault_search/schemas.py
@@ -17,23 +17,6 @@ if TYPE_CHECKING:
     from .indexer import VaultIndex
 
 # ---------------------------------------------------------------------------
-# Domain errors
-# ---------------------------------------------------------------------------
-
-
-class VaultSearchError(Exception):
-    """vault-search-mcp ドメインの基底例外."""
-
-
-class NoteNotFoundError(VaultSearchError):
-    """指定された path のノートがインデックスに存在しない."""
-
-    def __init__(self, path: str) -> None:
-        super().__init__(f"Note not found: {path}")
-        self.path = path
-
-
-# ---------------------------------------------------------------------------
 # Search responses
 # ---------------------------------------------------------------------------
